Split Organ Models from Devices; Change navbar and homepage
