+++ microdevices/templates/microdevices/microdevice_add.html
@@ -11,7 +11,7 @@
 {% endblock %}
 
 {% block breadcrumbs %}
-    <li><a href="/microdevices/">Devices &amp; Organ Models</a></li>
+    <li><a href="/microdevices/device">Devices</a></li>
     <li class="active">
         {% if object %}
             Edit <em>{{ object }}</em>
