+++ microdevices/templates/microdevices/microdevice_detail.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block breadcrumbs %}
-    <li><a href="/microdevices/">Devices &amp; Organ Models</a></li>
+    <li><a href="/microdevices/device">Devices</a></li>
     <li class="active">{{ object.device_name }}</li>
 {% endblock %}
