+++ microdevices/templates/microdevices/organmodel_add.html
@@ -2,7 +2,7 @@
 {% load static %}
 
 {% block breadcrumbs %}
-    <li><a href="/microdevices/">Devices &amp; Organ Models</a></li>
+    <li><a href="/microdevices/model">Organ Models</a></li>
     <li class="active">
         {% if object %}
             Edit <em>{{ object }}</em>
