+++ microdevices/templates/microdevices/organmodel_detail.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block breadcrumbs %}
-    <li><a href="/microdevices/">Devices &amp; Organ Models</a></li>
+    <li><a href="/microdevices/model">Organ Models</a></li>
     <li class="active">{{ object.model_name }}</li>
 {% endblock %}
