+++ mps/templates/base.html
@@ -221,10 +221,22 @@
           </li>
         </ul>
         </li>
+        <li title="Analyze bioactivity and drugtrial data" class="dropdown">
+        <a class="dropdown-toggle">Analysis<span class="caret"></span></a>
+        <ul class="dropdown-menu" role="menu">
+          <li title="Generate a heatmap for selected bioactivity data">
+              <a href="/bioactivities/heatmap/#filter">Heatmap</a>
+          </li>
+          <li title="Generate a dendrogram for selected compound/bioactivity data">
+              <a href="/bioactivities/cluster/#filter">Cluster</a>
+          </li>
+        </ul>
+        </li>
         <li title="View/add organ models and microdevices" class="dropdown">
         <a class="dropdown-toggle">Models<span class="caret"></span></a>
         <ul class="dropdown-menu" role="menu">
-          <li title="View organ models and microdevices"><a href="/microdevices/">View Models &amp; Devices</a></li>
+          <li title="View organ models"><a href="/microdevices/model">View Models</a></li>
+          <li title="View microdevices"><a href="/microdevices/device">View Devices</a></li>
           <li class="divider"></li>
           <li class="dropdown-header">Requires Permission</li>
           <li><a href="/microdevices/device/add/">Add Microdevice</a></li>
@@ -259,17 +271,6 @@
           <li><a href="/cellsamples/celltype/add/">Add Cell Type</a></li>
           <li><a href="/cellsamples/cellsubtype/add/">Add Cell Subtype</a></li>
           <li><a href="/cellsamples/cellsample/add/">Add Cell Sample</a></li>
-        </ul>
-        </li>
-        <li title="Analyze bioactivity and drugtrial data" class="dropdown">
-        <a class="dropdown-toggle">Analysis<span class="caret"></span></a>
-        <ul class="dropdown-menu" role="menu">
-          <li title="Generate a heatmap for selected bioactivity data">
-              <a href="/bioactivities/heatmap/#filter">Heatmap</a>
-          </li>
-          <li title="Generate a dendrogram for selected compound/bioactivity data">
-              <a href="/bioactivities/cluster/#filter">Cluster</a>
-          </li>
         </ul>
         </li>
         <li title="View help">
